+++ scripts/dmg-settings.py
@@ -1,0 +1,17 @@
+import os
+
+format = "UDZO"
+filesystem = "HFS+"
+files = [defines["app"], (defines["guide"], "Installation.txt")]
+symlinks = {"Applications": "/Applications"}
+background = defines["background"]
+window_rect = ((120, 120), (720, 460))
+default_view = "icon-view"
+show_toolbar = False
+show_status_bar = False
+show_sidebar = False
+show_pathbar = False
+include_icon_view_settings = True
+icon_size = 96
+text_size = 14
+icon_locations = {os.path.basename(defines["app"]): (190, 210), "Applications": (530, 210), "Installation.txt": (650, 402)}